+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# [1.10.0](https://github.com/eyobassey/RapidCapsuleMobile-/compare/v1.9.2...v1.10.0) (2026-03-20)
+
+### Features
+
+- **ui:** add medical disclaimer to all Dr Eka tabs and report views ([2baffb1](https://github.com/eyobassey/RapidCapsuleMobile-/commit/2baffb1369e7c1eddc8cc50a8bd84138ad7ed088))
+
+## [1.9.2](https://github.com/eyobassey/RapidCapsuleMobile-/compare/v1.9.1...v1.9.2) (2026-03-20)
+
+### Bug Fixes
+
+- parse health_score object in Dr Eka mobile and add notification routing ([0ce6f8e](https://github.com/eyobassey/RapidCapsuleMobile-/commit/0ce6f8e9659016a82d27222d2462d0ffaf1ece77))
+
+## [1.9.1](https://github.com/eyobassey/RapidCapsuleMobile-/compare/v1.9.0...v1.9.1) (2026-03-20)
+
+### Bug Fixes
+
+- guard against Invalid Date in Dr Eka reports and add null checks ([0575e57](https://github.com/eyobassey/RapidCapsuleMobile-/commit/0575e572bed8dc33c3547cd7c3765b8ea194d70e))
+
+# [1.9.0](https://github.com/eyobassey/RapidCapsuleMobile-/compare/v1.8.2...v1.9.0) (2026-03-20)
+
+### Features
+
+- add Dr. Eka AI physician screens to mobile app ([bf556e2](https://github.com/eyobassey/RapidCapsuleMobile-/commit/bf556e219222a94d1f053e610077ab8c5602df58))
+
 ## [1.8.2](https://github.com/eyobassey/RapidCapsuleMobile-/compare/v1.8.1...v1.8.2) (2026-03-20)
 
 ### Bug Fixes